updated to new version of ipimb part of LCLS

+++ LCLS/pdsdata/ipimb/src/ConfigV1.cc
@@ -17,8 +17,8 @@ ConfigV1::~ConfigV1() {}
 
 void ConfigV1::dump() const {
   printf("------Ipimb Config-------------\n");
-  printf("Trigger counter 0x%llx\n", (long long unsigned int)triggerCounter());
-  printf("Serial ID 0x%llx\n", (long long unsigned int)serialID());
+  printf("Trigger counter 0x%llx\n", triggerCounter());
+  printf("Serial ID 0x%llx\n", serialID());
   printf("Charge range %d\n", chargeAmpRange());
   printf("Reset length %ld ns\n", (unsigned long)resetLength());
   printf("Reset delay %d ns\n", resetDelay());
